+++ database/migrations/2018_01_25_140309_create_samples_table.php
@@ -14,9 +14,9 @@ class CreateSamplesTable extends Migration
     public function up()
     {
         Schema::create('samples', function (Blueprint $table) {
-            $table->bigIncrements('id');
-            $table->bigInteger('national_sample_id')->unsigned()->nullable()->index();
-            $table->bigInteger('patient_id')->unsigned()->index(); 
+            $table->increments('id');
+            $table->integer('national_sample_id')->unsigned()->nullable()->index();
+            $table->integer('patient_id')->unsigned()->index(); 
             // $table->bigInteger('batch_id')->unsigned()->index();
             $table->double('batch_id', 14, 2)->unsigned()->index();
             $table->tinyInteger('amrs_location')->nullable();
@@ -40,7 +40,7 @@ class CreateSamplesTable extends Migration
             $table->tinyInteger('spots')->unsigned()->nullable();
             $table->string('comments', 30)->nullable();
             $table->string('labcomment', 50)->nullable();
-            $table->bigInteger('parentid')->unsigned()->default(0)->nullable()->index();
+            $table->integer('parentid')->unsigned()->default(0)->nullable()->index();
             $table->tinyInteger('rejectedreason')->unsigned()->nullable();
             $table->tinyInteger('reason_for_repeat')->unsigned()->nullable();
             $table->string('interpretation', 100)->nullable();
